requirements.txt
absl-py==1.0.0
attrs==21.2.0
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Deprecated==1.2.13
dm-tree==0.1.6
filelock==3.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
gym==0.21.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.3
jsonschema==3.2.0
kiwisolver==1.3.1
lz4==3.1.10
Markdown==3.3.6
matplotlib==3.3.4
msgpack==1.0.3
networkx==2.5.1
numpy==1.19.5
numpy-stl==2.16.3
oauthlib==3.1.1
opencv-python==4.5.4.60
pandas==0.24.2
Pillow==8.4.0
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-utils==2.6.3
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
ray==1.7.1
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tifffile==2020.9.3
torch==1.9.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0

multiagent~=0.0.1
setuptools~=58.0.4
gfootball~=2.10.1
zlib~=1.2.11